authorisation is completed

+++ src/tasks/tasks.service.ts
@@ -17,12 +17,12 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) { }
 
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const task = await this.taskRepository.findOne({ where: { id, userId: user.id } });
         if (!task) {
             throw new NotFoundException(`Task with ID "${id}" was not found`);
         }
@@ -36,15 +36,18 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     };
 
-    async deleteTask(id: number): Promise<messageResponse> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<messageResponse> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" was not found`);
         }
